Update patch set 3
Patch Set 3:

> Has a bug.  Code extracted from recently-merged escape_param.go
 > misses a necessary moved-to-heap:
 > 
 > var sink interface{}
 > 
 > // &in -> out
 > func param11(k **int) ***int { // ERROR "leaking param: k to result
 > ~r1" "moved to heap: k$"
 > return &k // ERROR "&k escapes to heap$"
 > }
 > 
 > func caller11c(x int) {
 > i := x              // ERROR "moved to heap: i$"  // THIS ONE IS
 > MISSING, and yes it escapes.
 > p := &i             // ERROR "&i escapes to heap$" "moved to heap:
 > p$"
 > sink = *param11(&p) // ERROR "&p escapes to heap$" "\*param11\(&p\)
 > escapes to heap"
 > // **int passed, ***int returned, **int after dereference.
 > }

Yup. I remember one of my changes uncovered another latent bug in esc.c.
Try to apply the funcParam changes from here:
https://go-review.googlesource.com/#/c/4720/2/src/cmd/gc/esc.c
That is, remove funcParam and replace it with a flow through OIND.
A positive change in itself, but also it fixed some bug as far as I remember.

Patch-set: 3