Update patch set 3
Patch Set 3:

> > What do you think of the idea of leaving the assemblers alone for
 > > now, and just changing the linker?
 > 
 > Can we guarantee that code written in assembler doesn't need to
 > access global data via the GOT?  I guess it's OK for the assembler
 > in the runtime, but it seems a bit risky to me in general.

I don't mean that we won't have to change the assembler at all.  I mean that perhaps -buildmode is not the right option for the assembler.

In particular I can easily imagine that we will want to segregate packages on go install based on build options.  The linker options for that don't matter, only the compiler/assembler options.  It will be natural to segregate based on the compiler/assembler options.  But I think that several -buildmode options will require exactly the same behaviour from the compiler/assembler.  So it may make sense to use compiler/assembler options to select the code generation we want, rather than using a general option like -buildmode.

Patch-set: 3
Label: Code-Review=0