Update patch set 4
Patch Set 4:

Regarding the issue to be opened - I'd appreciate it if someone could take a look at my analysis of what is going on here and help me generalize it for an issue against the compiler. I'm not super confident in my assessment, and I don't think I know the compiler well enough to create a meaningful issue here on my own.

Patch-set: 4
Label: Code-Review=0