Update patch set 3
Patch Set 3:

> Yup. I remember one of my changes uncovered another latent bug in
 > esc.c.
 > Try to apply the funcParam changes from here:
 > https://go-review.googlesource.com/#/c/4720/2/src/cmd/gc/esc.c
 > That is, remove funcParam and replace it with a flow through OIND.
 > A positive change in itself, but also it fixed some bug as far as I
 > remember.

It looks like funcParam is indeed involved in this bug, so I am trying this change.
To be clear, the funcParam changes do NOT include changes involving either OCONVIFACE or ODOTTYPE, correct?

I'm trying to decide if you also intended for the level modifications to be included, since those do interact with the use of funcParam; I'll study this some more.

Patch-set: 3