Update patch set 3
Patch Set 3:

> Does deletion of funcParam alone fix the bug?

(repeat of accidental email reply)

It does not appear to, but I like those changes (they make more sense than the funcParam noodling) so I am going to work from them for a little while.  I suspect a glitch in the level arithmetic -- a level of indirection is added when k leaks out through param11 but I don't think that the "leak" captures that properly -- i.e., it leaks at -1, not zero, adding one with the * should take it to 0, not 1.  That's the hypothesis I'm testing now, at least.

Also, p should also be heap allocated, but (in small example) is not.  That's consistent with the hypothesis, and I think it lets me simplify the test case.

Patch-set: 3