Update patch set 1
Patch Set 1:

> I'm slightly concerned about there being no accounting of the
 > background sweep, but that's true before this CL too.

It looks to me like background sweeping will take almost no time away from running Gs (that is, it's practically just idle sweeping). If I'm right and it has basically no detriment to application performance, then I'm not sure there's value in reporting it to the user. (Of course, it may be useful to report to *us* in the "verbose" mode I propose.)

Patch-set: 1
Label: Code-Review=0