Update patch set 2
Patch Set 2:

Currently the @time part of the line prints the time that the GC cycle started.

Should this instead be when the GC cycle *ended*? That's when the trace line is printed, when the cumulative utilization that immediately follows the @time applies to, and the time of the final heap size. It would also give us room to add background sweep information if we wanted to; otherwise its stats would be before the @ time, while all other stats would be after the @ time.

Patch-set: 2
Label: Code-Review=0