Update patch set 4
Patch Set 4:

> (4 comments)
 > 
 > > (3 comments)
 > >
 > > Nothing jumps out at me as a major problem for garbage
 > collection.
 > > For the most part, the garbage collector just cares about heap vs
 > > non-heap pointer, which this has no effect on.
 > 
 > Oops, I guess I was super unspecific about this. The parts I was a
 > little uneasy about were the loops I've added markroot (like, is it
 > important that it only invokes scanblocks once?) and vague worries
 > about performance -- I can see that the performance of findfunc is
 > going to be somewhat important, but I don't know if it's "every
 > extra instruction hurts" hot.  I can think of ways to make it
 > faster (a tree-type structure probably) but I'd just as soon not :)

The markroot change should fine. It wouldn't be good to invoke scanblock multiple times on the same memory, but calling it multiple times on disjoint regions of memory is fine.

I'm less sure about the performance requirements of findfunc, so I'm afraid I can't advise there.

If we keep the moduledata in a slice instead of a linked list and keep that sorted by minpc, then findmoduledatap could use a binary search. No trees required. :)

 > > BTW, you might wan to use the -topic argument to git mail to give
 > > all of the patches in the series a Gerrit topic to tie them
 > > together. I think if you do this the next time you mail an update
 > > it will apply to all updated changes (though you might have to
 > > hand-topic some of the earlier ones if their commit hashes
 > haven't
 > > changed).
 > 
 > Oh yes, very good point (gerrit newbie here).  Will do that next
 > time I push changes.

This isn't particularly well documented (for that matter, nothing about multi-change workflow is well documented).

Patch-set: 4