Update patch set 1
Abandoned

It turns out that this was only required for the most terribly obscure reasons.

Basically, because of the normalization of interpunct to period when writing out symbols, 6l -linkshared was not loading the $func·f symbols from the DSO but rather ended up creating additional copies of the descriptors in the executable -- which had relocs that filled them out with PLT stubs. For other reasons, I disabled the symbol name normalization and now 6l -linkshared creates an executable that references the descriptor in the DSO and this contains a pointer to the function itself, not the stub.

Patch-set: 1
Status: abandoned