Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

At some point, can you write up a separate short note on how to handle types in shared libraries and send it to golang-dev?  It would be nice to get the scheme written down somewhere (not that the old scheme is written down).

Patch-set: 4
Label: Code-Review=+1