Update patch set 1
Patch Set 1:

They already have to be in sync for most of the struct.  Is this really worth doing?

If you like, we could add a dynamic check that they are the same by having one of the fields by the size of the struct.  The linker could initialize that field to its notion of the size.  A runtime package init function could compare that to the size that it wants, and throw if it is different.

Patch-set: 1
Label: Code-Review=0