Update patch set 1
Patch Set 1:

Sending this for discussion. It is untested, as I do not have arm64 or ppc hardware. I'd love testing and benchmarking help!

Also, it seems like the proper type for len and cap is TINT, not TUINT, but this just brings 7g and 9g in line with what the other architectures have had for a long time.

If there's interest in changing that, I'm happy to do it in another CL.

Patch-set: 1
Label: Code-Review=0