Update patch set 3
Patch Set 3:

When a program is linked against shared libraries, the shared library initialization will happen before any of the initialization in the main program.  When the Go runtime is in the main program, as is always the case today, shared library initialization will happen before the Go runtime initialization.

Patch-set: 3
Label: Code-Review=0