cmd/go: accept -nodeps flag for go get
When fetching a package with known, fixed
dependencies, it is useful to be able to download
the package only without its dependencies.

With the current behaviour, unwanted dependencies can be
fetched which is problematic for tools that need to
check that all dependencies are correctly versioned.

The problem scenario looks something like this:

	fetch required package from VCS
	for each dependency {
		go get -d dependency
		update dependency to required version
	}

Without the -nodeps flag to go get, the "go get"
line will fetch the dependencies from the tip of the
dependency repo, rather than from the required
version.

An alternative might be to entirely fork the package-downloading
part of the Go tool, allowing the tool to accomplish
the same thing directly, or to factor it out into its own
package.

Change-Id: I08fc3c60eca0bdf475244b313850826a5a56a777