Update patch set 2
Patch Set 2:

> Dmitriy? Any comments? Otherwise fine to submit.

This is waiting on me.

I'm planning to return to this when I'm done with the pacer and try implementing Dmitry's proposed approach, which I think will be cleaner.

The part around entering the mark phase is going to be non-trivial to merge with the pacer anyway, and I'd rather get the pacer in first (or, at least, up for review).

Patch-set: 2
Label: Code-Review=0