Update patch set 2
Patch Set 2:

I added a test case to the issue. Based on the reproduction, I'm not sure that this is the right fix. The reproduction requires two runs, which suggests that the problem is that something is not getting cleared appropriately when being returned to a free list. (I haven't looked closely, though, so I could be wrong.)

Patch-set: 2
Label: Code-Review=0