os, internal/syscall: use Fstatat for Readdir on Linux and BSDs
The existing implementation of Readdir on unix uses Lstat and the path
of file (f.name). This is incorrect when the directory is moved or
the path is not walkable at the moment. Windows uses the handle
for readdir and thus works like Fstatat.

Old behaviour remains on darwin and solaris. Windows and Plan9
are not affected.

Change-Id: I4a0b1d7e967138e6cb9c587066b8007fcc318753