Update patch set 2
Patch Set 2:

Just as a note: this can make the issue described in the last part of this comment worse:

// When a goroutine needs to allocate large-object span from heap,
// it sweeps spans until it frees at least that many pages into heap. There is
// one case where this may not suffice: if a goroutine sweeps and frees two
// nonadjacent one-page spans to the heap, it will allocate a new two-page
// span, but there can still be other one-page unswept spans which could be
// combined into a two-page span.


I don't know how important it is. I never seen a case where this issue fires.
Maybe I used +1 to account for this issue -- if we resort to small span sweeping (which is the last resort), then we want to over-sweep. Or maybe it was just a bug.

Patch-set: 2