Update patch set 3
Patch Set 3:

Predictably, in retrospect, this hosed the darwin-arm builder. It got stuck in an odd state which tricked me into upgrading, which is now triggering a bug in ios-deploy.

Underlying binaries still run if I hack up ios-deploy, but doing so breaks the test harness. So I'm putting this off for a few days in the hopes that some intrepid person working on ios-deploy fixes it for iOS 8.3.

Until then we are without a darwin/arm builder.

Patch-set: 3