Update patch set 2
Patch Set 2:

> I wonder also whether we should use TINT instead of TUINT here.  But that can be for another change.

On second thought, I'd just as soon not touch it at all until we have to. I'm nervous about the unexpected consequences. This is fragile stuff, as we can see.


> Why was this code put in in the first place?

I traced it back to https://golang.org/cl/123170043, which was the very first powerpc commit. I imagine that Minux just did the sensible thing, which turned out to be wrong even though it was right. It then got copied to 7g, along with everything else from 9g--no obvious reason to change it.

Patch-set: 2
Label: Code-Review=0