Update patch set 1
Patch Set 1: Code-Review-1

I think the first thing to do is to get the tests to pass.

Not clear to me this is the right place to manage the issue of new packages. It makes the test very expensive and, as minux says, the approach isn't sound anyway.

Patch-set: 1
Label: Code-Review=-1