Update patch set 1
Patch Set 1:

It seems like this patch will still fail some requests that should succeed. Consider a non-idempotent request issued on a connection that has *already* received the EOF from the server. The higher layers would only detect the EOF after making the request and would not retry due to the safety rules. While this diff improves reliability for idempotent requests, it does not address the underlying issue of reusing connections that are already dead. We need to check the connection for EOF *before* using it for the real request to fix this problem.

Patch-set: 1