Update patch set 7
Patch Set 7:

> This should have a test under misc/cgo, probably in a new
 > directory.  Have a global constructor set up a signal handler for
 > SIGSEGV.  Test that Go code handles nil dereferences via a runtime
 > panic, and that C code invokes the signal handler.

Added the tests.

Two open issues are:
  (1) do we report forwarded signals to os/signal?
  (2) do we forward only SIGSEGV, SIGILL, and SIGBUS in executable mode?

If we want to do (1), I presume that we should emulate the behavior of non-forwarded signals.  I tried to implement that but the code gets a bit complex, just a warning.

For (2), I presume we'd still forward all signals in library/archive modes?  Java apparently uses a bunch of other signals, not sure how important it is to forward them.

Patch-set: 7
Label: Code-Review=0