Update patch set 8
Patch Set 8:

> (1) do we report forwarded signals to os/signal?
 >
 > (2) do we forward only SIGSEGV, SIGILL, and SIGBUS in executable
 > mode?
 > 
 > If we want to do (1), I presume that we should emulate the behavior
 > of non-forwarded signals.  I tried to implement that but the code
 > gets a bit complex, just a warning.
 > 
 > For (2), I presume we'd still forward all signals in
 > library/archive modes?  Java apparently uses a bunch of other
 > signals, not sure how important it is to forward them.

I now think that we should only forward the synchronous signals, namely the ones for which the _SigPanic flag is set (and we should only forward when the context sigcode != SI_USER).

The argument is, basically, SIGPROF and SIGALRM, plus signals sent by syscall.Kill and syscall.Tgkill.  Those signals do not depend on what code is executing.  It was a mistake on my part to think that the code being executed is a concern for all signals.  It's only a concern for a few signals that are triggered directly by code execution.

Forwarding the synchronous signals should be all that is required to get basic JVM support working.

This leaves open the issue of how to share other signals with other runtimes, but forwarding based on what code is executing can not be the answer.  So lets leave that issue for a later day.

So that answers question 2: we always forward the synchronous signals, and we never forward the non-synchronous signals.

And I think that obviates question 1: the synchronous signals are never passed to os/signal anyhow.

Patch-set: 8
Label: Code-Review=0