Update patch set 1
Patch Set 1:

Can you elaborate about "it was a win for 6g"? Losing a register seems like a strict loss, and the $0 only costs one byte in the instruction encoding. For two moves it's break even in code size, and for more moves, $0 is slightly bigger but still not bad.

Patch-set: 1