net/http: Fix race between CancelRequest and request EOF
The bodyEOFSignal function would release the persistConn back to the
transport before the readLoop cleared the ReqCanceler. If a request was
canceled after the client was done reading the body of the response,
the cancellation could end up closing the persistConn with no pending
requests. When the next request then came along, it would find the
persistConn dead and error out.

Fix by pulling the setReqCanceler into the bodyEOFSignal function
before we give the connection back to the transport

Fixes #10474

Change-Id: Ib14cf920263d56231a5aa6bb435b9679210bbc43