Update patch set 1
Patch Set 1:

There is another race in this code:
https://github.com/golang/go/issues/9946

I outlined the problem here:
https://github.com/golang/go/issues/9946#issuecomment-75291782

Brad tried to fix it here:
https://go-review.googlesource.com/#/c/8550/
but it does not look like a proper solution to me.

Since you fix a race in the very same code, it can make sense to fix both of them.

Patch-set: 1
Label: Code-Review=0