cmd/link, cmd/internal/goobj: update constants, regenerate testdata
The constants in cmd/internal/goobj had gone stale (we have three copies of
these constants now, working on reducing that was what got me to noticing this).

I haven't inspected the changes to link.hello.darwin.amd64 deeply, but the path
is clearly wrong -- is asm -trimpath not doing what it should?

Change-Id: Iaa1511f458a32228c2df2ccd0076bb9ae212a035