Update patch set 5
Patch Set 5:

I am leaning towards only having a _GCmark phase and getting
rid of the _GCmarksetup phase. We can add another help global to  indicate that the assists are appropriate. This would not have to have a barrier since if the G does not help it is not a correctness issue.

Patch-set: 5
Label: Code-Review=0