Update patch set 1
Patch Set 1:

> I remember the benchmarks not moving. The commit message from CL
 > 2606 says this reduces godoc binary size by 0.29%.
 > 
 > Right now, building the stdlib shows ~67% two moves, ~28% three
 > moves, and the rest incidental. It also shows 75 instances of zero
 > moves (?).
 > 
 > By all means leave the TODO if you want to dig more. :)

FWIW, storing $0 in a register and then moving that register into memory is directly at odds with propagation algorithms. For example if the final memory location is itself registerized, then a direct MOV of $0 into the now-registerized value can turn into a simple register clear, while MOV of a $0 through a different register is probably not going to propagate as well, if at all.

Also when the initializations disappear entirely because the values are registerized and those initializations are not read before subsequent writes, the MOV $0, BX to prime the register is left behind and not eliminated as well.

Not going to touch anything in this CL, and Keith's work will make this irrelevant, but it may be possible to improve in the short term.

Patch-set: 1
Label: Code-Review=0