Update patch set 1
Patch Set 1:

> A couple of foolish questions:
 > 
 > Could you give an example of a non-standard compilation option? 

See the values for codegenArg in buildmodeInit in cmd/go/build.go.  Because of these options, the objects have to be recompiled (unless you use go install, of course).

 > Why does it only apply when p.cmdline is true?
 
It's only necessary to add the import to the final link, which is the package mentioned on the command line.

 > Does this same reasoning apply to c-archive?

No, because the linker doesn't force external linking mode for c-archive.  The relevant code is in archinit in cmd/6l/obj.go.

Patch-set: 1