Update patch set 1
Patch Set 1:

> I don't understand how an idle P which can't find a runnable G is
 > told to work on behalf of the GC. I expected to see the logic in
 > findrunnable but didn't. Any clues?

If a P is running and calls findrunnable, the last thing findrunnable does before marking its P idle and releasing it is to check if it should do marking.

If a P is *already* idle, things are subtler. I think we might actually wake up idle Ps as a side-effect of resetspinning() calling wakep(), but I haven't convinced myself of this yet.

Patch-set: 1