Update patch set 1
Patch Set 1:

> > > Why does it only apply when p.cmdline is true?
 > >
 > > It's only necessary to add the import to the final link, which is
 > > the package mentioned on the command line.
 > 
 > It looks like p.cmdline is only set for packages created from
 > listed .go files. Doesn't this logic apply to any way of specifying
 > a package?

You're right, I was confused.  I changed it to check for a package name of "main", which is what is used to set a.link = true in builder.action.

 > > > Does this same reasoning apply to c-archive?
 > >
 > > No, because the linker doesn't force external linking mode for
 > > c-archive.  The relevant code is in archinit in cmd/6l/obj.go.
 > 
 > OK. Not relevant to this CL, but I suspect c-archive doesn't work
 > yet with internal linking. I thought external was forced, but I
 > agree there's no code to do it.

I don't see any particular need at present to force external linking for c-archive.  For c-shared we need to force it because the Go linker can't generate a correct shared library.  For c-archive all we need is an object file, and the Go linker does know how to generate that.

Patch-set: 1
Label: Code-Review=0