net: add RemoteAddr field to OpError
Not only by network, transport-layer intermediaries but by
virtualization stuff in a node makes things hard to identify the root
cause of weird faults after disaster happened. This change adds
RemoteAddr field to OpError to be able to represent a 5-tuple of
internet transport protocols for helping dealing with complicated
systems.

Updates #4856.

Change-Id: I96a523fe391ed14406bfb21604c461d4aac2fa19