Update patch set 10
Patch Set 10: Run-TryBot+1

This is the patch that was just measured with compilation of std.

It includes calculation of a new quantity, 
"maximum level of a copy-headed suffix before &" (phrasing TBD)
which captures+generalizes the way that * cancels a following &.

There's still work needed on this, in particular field names and comments,
and most likely some style issues as well.

Patch-set: 10
Label: Run-TryBot=+1