Update patch set 7
Patch Set 7: -Run-TryBot

The amd64-linux and nacl-386 failures are not related to this CL.  These two have failed on a different CL of mine, which later passed on Go Build Dashboard:
https://storage.googleapis.com/go-build-log/1832a749/linux-amd64-race_92eea9c8.log
https://storage.googleapis.com/go-build-log/1832a749/nacl-386_f6344d61.log

The openbsd-386 failure is the same OOM failure as linux-amd64, so it's probably also not related.

This CL is good to be cherry-picked.

Patch-set: 7
Reviewer: Gerrit User 6605 <6605@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=0