runtime: panic when idling a P with runnable Gs
This adds a check that we never put a P on the idle list when it has
work on its local run queue.

This seems like the most likely cause of the "spinning with local
work" failures that have been happening on the dashboard. Hopefully
this will help debug these by failing faster.

Change-Id: Ifcfab750de60c335148a7f513d4eef17be03b6a7